refactor: rename EncodeProfile to Recipe
