--- src/Service.Dashboard/src/Application.Contracts/EncodeProfiles/EncodeProfileDelete.cs
@@ -1,26 +1,0 @@
-using FluentValidation;
-using Giantnodes.Infrastructure.Messages;
-
-namespace Giantnodes.Service.Dashboard.Application.Contracts.EncodeProfiles;
-
-public sealed class EncodeProfileDelete
-{
-    public sealed record Command : Message
-    {
-        public required Guid Id { get; init; }
-    }
-
-    public sealed class Validator : AbstractValidator<Command>
-    {
-        public Validator()
-        {
-            RuleFor(p => p.Id)
-                .NotEmpty();
-        }
-    }
-
-    public sealed record Result
-    {
-        public required Guid Id { get; init; }
-    }
-}
