--- src/Service.Dashboard/src/Domain/Aggregates/EncodeProfiles/Repositories/IEncodeProfileRepository.cs
@@ -1,7 +1,0 @@
-using Giantnodes.Infrastructure.Domain.Repositories;
-
-namespace Giantnodes.Service.Dashboard.Domain.Aggregates.EncodeProfiles.Repositories;
-
-public interface IEncodeProfileRepository : IRepository<EncodeProfile>
-{
-}
