--- src/Service.Dashboard/src/Domain/Aggregates/EncodeProfiles/Specifications/IsEncodableSpecification.cs
@@ -1,11 +1,0 @@
-using Giantnodes.Infrastructure.Domain.Specifications.Compositions;
-
-namespace Giantnodes.Service.Dashboard.Domain.Aggregates.EncodeProfiles.Specifications;
-
-public class IsEncodableSpecification : AndSpecification<EncodeProfile>
-{
-    public IsEncodableSpecification()
-        : base(new IsQualitySupportedSpecification(), new IsTuneSupportedSpecification())
-    {
-    }
-}
