--- src/Service.Dashboard/src/Domain/Aggregates/EncodeProfiles/Specifications/IsQualitySupportedSpecification.cs
@@ -1,13 +1,0 @@
-using System.Linq.Expressions;
-using Giantnodes.Infrastructure.Domain.Specifications;
-
-namespace Giantnodes.Service.Dashboard.Domain.Aggregates.EncodeProfiles.Specifications;
-
-public class IsQualitySupportedSpecification : Specification<EncodeProfile>
-{
-    public override Expression<Func<EncodeProfile, bool>> ToExpression()
-    {
-        return x => x.Quality == null ||
-                    (x.Quality.Value >= x.Codec.Quality.Min && x.Quality.Value <= x.Codec.Quality.Max);
-    }
-}
