+++ src/Service.Dashboard/src/Domain/Aggregates/Recipes/Repositories/IRecipeRepository.cs
@@ -1,0 +1,7 @@
+using Giantnodes.Infrastructure.Domain.Repositories;
+
+namespace Giantnodes.Service.Dashboard.Domain.Aggregates.Recipes.Repositories;
+
+public interface IRecipeRepository : IRepository<Recipe>
+{
+}
