+++ src/Service.Dashboard/src/Domain/Aggregates/Recipes/Specifications/IsQualitySupportedSpecification.cs
@@ -1,0 +1,13 @@
+using System.Linq.Expressions;
+using Giantnodes.Infrastructure.Domain.Specifications;
+
+namespace Giantnodes.Service.Dashboard.Domain.Aggregates.Recipes.Specifications;
+
+public class IsQualitySupportedSpecification : Specification<Recipe>
+{
+    public override Expression<Func<Recipe, bool>> ToExpression()
+    {
+        return x => x.Quality == null ||
+                    (x.Quality.Value >= x.Codec.Quality.Min && x.Quality.Value <= x.Codec.Quality.Max);
+    }
+}
