+++ src/Service.Dashboard/src/HttpApi/Resolvers/Recipes/Queries/RecipeFindOne.cs
@@ -1,0 +1,18 @@
+using Giantnodes.Service.Dashboard.Domain.Aggregates.Recipes;
+using Giantnodes.Service.Dashboard.Persistence.DbContexts;
+using Microsoft.EntityFrameworkCore;
+
+namespace Giantnodes.Service.Dashboard.HttpApi.Resolvers.Recipes.Queries;
+
+[ExtendObjectType(OperationTypeNames.Query)]
+public class RecipeFindOne
+{
+    [UseFirstOrDefault]
+    [UseProjection]
+    [UseFiltering]
+    [UseSorting]
+    public IQueryable<Recipe> Recipe([Service] ApplicationDbContext database)
+    {
+        return database.Recipes.AsNoTracking();
+    }
+}
