+++ src/Service.Dashboard/src/Infrastructure/Repositories/RecipeRepository.cs
@@ -1,65 +1,62 @@
 using System.Linq.Expressions;
-using Giantnodes.Service.Dashboard.Domain.Aggregates.Encodes;
-using Giantnodes.Service.Dashboard.Domain.Aggregates.Encodes.Repositories;
+using Giantnodes.Service.Dashboard.Domain.Aggregates.Recipes;
+using Giantnodes.Service.Dashboard.Domain.Aggregates.Recipes.Repositories;
 using Giantnodes.Service.Dashboard.Persistence.DbContexts;
 using Microsoft.EntityFrameworkCore;
 
 namespace Giantnodes.Service.Dashboard.Infrastructure.Repositories;
 
-public class EncodeRepository : IEncodeRepository
+public class RecipeRepository : IRecipeRepository
 {
     private readonly ApplicationDbContext _database;
 
-    public EncodeRepository(ApplicationDbContext database)
+    public RecipeRepository(ApplicationDbContext database)
     {
         _database = database;
     }
 
-    public IQueryable<Encode> ToQueryable()
+    public IQueryable<Recipe> ToQueryable()
     {
         return _database
-            .Encodes
-            .Include(x => x.File)
-            .Include(x => x.Recipe)
-            .Include(x => x.Snapshots)
+            .Recipes
             .AsQueryable();
     }
 
     public Task<bool> ExistsAsync(
-        Expression<Func<Encode, bool>> predicate,
+        Expression<Func<Recipe, bool>> predicate,
         CancellationToken cancellation = default)
     {
         return ToQueryable().AnyAsync(predicate, cancellation);
     }
 
-    public Task<Encode> SingleAsync(
-        Expression<Func<Encode, bool>> predicate,
+    public Task<Recipe> SingleAsync(
+        Expression<Func<Recipe, bool>> predicate,
         CancellationToken cancellation = default)
     {
         return ToQueryable().SingleAsync(predicate, cancellation);
     }
 
-    public Task<Encode?> SingleOrDefaultAsync(
-        Expression<Func<Encode, bool>> predicate,
+    public Task<Recipe?> SingleOrDefaultAsync(
+        Expression<Func<Recipe, bool>> predicate,
         CancellationToken cancellation = default)
     {
         return ToQueryable().SingleOrDefaultAsync(predicate, cancellation);
     }
 
-    public Task<List<Encode>> ToListAsync(
-        Expression<Func<Encode, bool>> predicate,
+    public Task<List<Recipe>> ToListAsync(
+        Expression<Func<Recipe, bool>> predicate,
         CancellationToken cancellation = default)
     {
         return ToQueryable().Where(predicate).ToListAsync(cancellation);
     }
 
-    public Encode Create(Encode entity)
+    public Recipe Create(Recipe entity)
     {
-        return _database.Encodes.Add(entity).Entity;
+        return _database.Recipes.Add(entity).Entity;
     }
 
-    public Encode Delete(Encode entity)
+    public Recipe Delete(Recipe entity)
     {
-        return _database.Encodes.Remove(entity).Entity;
+        return _database.Recipes.Remove(entity).Entity;
     }
 }
